[REFACTOR][PYTHON] Move tvm.contrib.nvcc to tvm.support.nvcc
tvm.contrib.nvcc is the NVCC/NVRTC compile driver. It is side-effect-imported
from python/tvm/__init__.py to register tvm_callback_cuda_compile and related
FFI callbacks at TVM startup, making it load-bearing for every CUDA codegen
flow. Moving it to tvm.support reflects its non-optional role. All callers
updated; tvm/__init__.py now imports from .support instead of .contrib. FFI
registry keys (tvm.contrib.nvcc.*) are left intact — renaming them is a
separate follow-up.
